push utility functions to Stringfromenvvar
