+++ libs/datamodel/core/src/configuration.rs
@@ -2,8 +2,10 @@ pub mod preview_features;
 
 mod configuration;
 mod datasource;
+mod env_vars;
 mod generator;
 
 pub use configuration::*;
 pub use datasource::*;
+pub use env_vars::*;
 pub use generator::*;
