+++ libs/datamodel/core/src/configuration/datasource.rs
@@ -1,7 +1,6 @@
-use std::{collections::BTreeMap, path::Path};
-
+use crate::StringFromEnvVar;
 use datamodel_connector::{Connector, ConnectorCapabilities};
-use serde::Serialize;
+use std::{collections::BTreeMap, path::Path};
 use url::Url;
 
 /// a `datasource` from the prisma schema.
@@ -30,6 +29,7 @@ impl std::fmt::Debug for Datasource {
             .field("url", &self.url)
             .field("documentation", &self.documentation)
             .field("active_connector", &&"...")
+            .field("shadow_database_url", &self.shadow_database_url)
             .finish()
     }
 }
@@ -98,12 +98,4 @@ impl Datasource {
             }
         }
     }
-}
-
-#[serde(rename_all = "camelCase")]
-#[derive(Clone, Debug, Serialize, PartialEq)]
-pub struct StringFromEnvVar {
-    /// contains the name of env var if the value was read from one
-    pub from_env_var: Option<String>,
-    pub value: String,
 }
