+++ libs/datamodel/core/src/transform/dml_to_ast/datasource_serializer.rs
@@ -20,15 +20,7 @@ impl DatasourceSerializer {
     fn lower_datasource(source: &Datasource) -> ast::SourceConfig {
         let mut arguments: Vec<ast::Argument> = vec![ast::Argument::new_string("provider", &source.active_provider)];
 
-        match source.url.from_env_var {
-            Some(ref env_var) => {
-                let values = vec![ast::Expression::StringValue(env_var.to_string(), ast::Span::empty())];
-                arguments.push(ast::Argument::new_function("url", "env", values));
-            }
-            None => {
-                arguments.push(ast::Argument::new_string("url", &source.url.value));
-            }
-        }
+        arguments.push(source.url.to_arg());
 
         ast::SourceConfig {
             name: ast::Identifier::new(&source.name),
