+++ libs/datamodel/core/tests/config/generators.rs
@@ -168,11 +168,13 @@ fn nice_error_for_unknown_generator_preview_feature() {
 fn retain_env_var_definitions_in_generator_block() {
     std::env::set_var("PROVIDER", "postgres");
     std::env::set_var("OUTPUT", "~/home/prisma/");
+    std::env::set_var("CUSTOM", "whatever");
 
     let schema1 = r#"
     generator js1 {
         provider = env("PROVIDER")
         output = env("OUTPUT")
+        custom = env("CUSTOM")
     }
     "#;
 
@@ -183,7 +185,7 @@ fn retain_env_var_definitions_in_generator_block() {
     "output": "~/home/prisma/",
     "binaryTargets": [],
     "previewFeatures": [],
-    "config": {}
+    "config": {"custom": "env"}
   }
 ]"#;
