+++ libs/datamodel/core/tests/config/mod.rs
@@ -1,4 +1,4 @@
-pub mod datasources_to_mcf;
+pub mod datasources;
 pub mod generators;
 mod nice_warnings;
 pub mod sources;
